Update keystone-operator to use common functions
To make different service operators can look similar
this moves implementation of the different required
resource to lib-common with interfaces that they
could be consumed easily. This demonstrates how it
could be done.

Also:
- service customization via CustomServiceConfig param
which gets added to the VM and inside the pod as
keystone/conf.d/custom.conf
- Nodeselector
- pod affinity - if possible two pods of the same
service should not run on the same worker node. If this
is not possible	the get still created on the same worker
node.

TODOs:
- finish conditions, right now only in db create
- set finalizers on the different created resources
to prevent deletion
- poc update to see if the dedicated update reconcile
func is required
- poc upgrade
- implement DefaultConfigOverwrite

Depends-On: